Revert "up"
This reverts commit 5b28dd6be97f4d48b891b03bee409d533a726589, reversing
changes made to 295d2bd868a80614503fd1060f12cf78d75e937e.

+++ AuctionKoi/src/main/java/uth/edu/auctionkoi/pojo/Bid.java
@@ -2,8 +2,6 @@ package uth.edu.auctionkoi.pojo;
 
 import jakarta.persistence.*;
 import lombok.Data;
-
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Entity
@@ -21,7 +19,7 @@ public class Bid {
     @JoinColumn(name = "user_id")
     private User user;
 
-    private BigDecimal bidAmount; // Giá người mua đặt
+    private Double bidAmount; // Giá người mua đặt
     private LocalDateTime bidTime; // Thời điểm đặt giá
 }
